author filter added

+++ admin/includes/view_all_post.php
@@ -103,7 +103,7 @@ if (isset($_POST['checkBoxArray'])) {
                 echo "<td>$post_comments</td>";
                 echo "<td>$post_date</td>";
                 echo "<td><a href='./posts.php?source=edit_post&p_edit={$post_id}'>Edit</td>";
-                echo "<td><a href='./posts.php?delete={$post_id}'>Delete</td>";
+                echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delte?'); \" href='./posts.php?delete={$post_id}'>Delete</td>";
                 echo "</tr>";
             }
